Move Transaction and related types to separate file
